弄好了输出

+++ Ai/ILayer.h
@@ -9,7 +9,7 @@ __interface ILayer
 	size_t GetInputSize() const;
 	size_t GetNeuralCount() const;
 	FloatingPoint Train(const ISample& sample);
-	_STD unique_ptr<INeuralProxy> GetNeural(int index);
+	_STD unique_ptr<INeuralProxy> GetNeural(size_t index);
 	void Transform(const _STD vector<FloatingPoint>& input, _Out_ _STD vector<FloatingPoint>& output) const;
 };
